Runner Install

+++ Gitlab-Docker-Install/docker-install.sh
@@ -3,9 +3,12 @@
 # Add Docker's official GPG key:
 sudo apt-get update
 sudo apt-get install ca-certificates curl
-sudo install -m 0755 -d /etc/apt/keyrings
-sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-sudo chmod a+r /etc/apt/keyrings/docker.asc
+sudo install -m 0755 -d /etc/apt/keyrings 
+
+if [ ! -f /etc/apt/keyrings/docker.asc ]; then
+    sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
+fi
 
 # Add the repository to Apt sources:
 echo \
@@ -20,3 +23,23 @@ sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin 
 
 # Docker service status info: 
 sudo systemctl status docker
+# Enable and start Docker service
+echo "Checking the status of the Docker service..."
+
+# Check if the Docker service is enabled
+if ! systemctl is-enabled --quiet docker; then
+    echo "Enabling Docker service..."
+    sudo systemctl enable docker
+else
+    echo "Docker service is already enabled."
+fi
+
+# Check if the Docker service is active
+if ! systemctl is-active --quiet docker; then
+    echo "Starting Docker service..."
+    sudo systemctl start docker
+else
+    echo "Docker service is already running."
+fi
+
+echo "Installation completed."
